# Conflicts:
#	extern/libogg/libogg-1.3.2-vc64.zip
#	extern/libtheora/libtheora-1.1.1-vc64.zip
#	extern/libvorbis/libvorbis-1.3.5-vc64.zip
#	extern/zlib/zlib-1.2.13-vc64.zip

+++ src/deigde/editors/conversation/src/gui/ceCanvasRuleOfThirdsAid.cpp
@@ -176,7 +176,7 @@ void ceCanvasRuleOfThirdsAid::pAddRect( int x1, int y1, int x2, int y2, const de
 	canvasPaint->SetThickness( 0 );
 	canvasPaint->SetLineColor( decColor( color, 0.0f ) /*color*/ );
 	canvasPaint->SetFillColor( color );
-	canvasPaint->SetOrder( pCanvasView->GetCanvasCount() );
+	canvasPaint->SetOrder( ( float )pCanvasView->GetCanvasCount() );
 	canvasPaint->SetTransparency( transparency );
 	canvasPaint->SetPosition( decPoint( x1, y1 ) );
 	canvasPaint->SetSize( decPoint( x2 - x1, y2 - y1 ) );
